requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Geospatial
folium>=0.14.0
shapely>=2.0.0
geopandas>=0.13.0

# Google Cloud (for BigQuery)
google-cloud-bigquery>=3.10.0
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0
db-dtypes>=1.1.0

# Utilities
python-dotenv>=1.0.0
pytz>=2023.3
duckdb>=0.10.0

# Optimization
scipy>=1.11.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0

# Development (optional)
pytest>=7.3.0
black>=23.3.0
